requirements.txt
openai>=1.0.0
python-dotenv
python-telegram-bot==20.7